+++ core/assemble/assemble.c
