+++ core/assemble/assemble.h
@@ -1,5 +1,5 @@
-#ifndef INIT_ASSEMBLE_ASSEMBLE_H
-#define INIT_ASSEMBLE_ASSEMBLE_H
+#ifndef __CORE_ASSEMBLE_ASSEMBLE_H
+#define __CORE_ASSEMBLE_ASSEMBLE_H
 
 /*
 # OPCODES
