+++ core/node/node.c
