+++ core/parser/parser.h
@@ -1,0 +1,7 @@
+#ifndef __CORE_PARSER_PARSER_H__
+#define __CORE_PARSER_PARSER_H__
+
+#include "tokenizer.h"
+
+void parse(TokenNode* tk);
+#endif
