+++ core/util/util.c
