+++ core/vm/vm.c
