--- init/parser/parser.h
@@ -1,7 +1,0 @@
-#ifndef __INIT_PARSER_H__
-#define __INIT_PARSER_H__
-
-#include "tokenizer.h"
-
-void parse(TokenNode* tk);
-#endif
