--- out2.hex
@@ -1,6 +1,0 @@
-0305
-0900
-3301
-2900
-3301
-B007
